+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Updated YACLIL function options
 - Now `subcommands` is an object, not array
 - Added `yaclil/types` path
+- Added padding for subcommands and options in help
 
 ## 1.4.7 2024-06-18